unipune/demo.py
from twilio.rest import Client

# Your Account SID from twilio.com/console
account_sid = "AC0e5bfb488cbe8bc41bdfb9b88ede331b"
# Your Auth Token from twilio.com/console
auth_token  = "2f897aea9013c8e75534f50767b362e9"

client = Client(account_sid, auth_token)

message = client.messages.create(
    to="+919270088467", 
    from_="+919270088467",
    body="Hello from Python!")

print(message.sid)
