requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==2.1.2
django-cors-headers==2.4.0
django-rest-knox==3.4.0
djangorestframework==3.9.0
djangotoolbox==1.8.0
idna==2.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.19
pyOpenSSL==18.0.0
python-http-client==3.1.0
pytz==2018.6
sendgrid==5.6.0
six==1.11.0
